finish the layers(domain, external, infra) need to do the integrate the presenter layer with the other layers
