--- lib/app/repositories/infra/datasources/git_repo_datasource.dart
@@ -1,3 +1,0 @@
-abstract class IGitRepoInfraDatasource {
-  Future<List> getRepos();
-}
