+++ lib/app/repositories/infra/datasources/gitrepo_datasource.dart
@@ -1,0 +1,3 @@
+abstract class IGitRepoDatasource {
+  Future<List> getRepos();
+}
